requirements.txt
Flask==3.1.3
Flask-Cors==6.0.2
Flask-SQLAlchemy==3.1.1
gunicorn==25.3.0
pandas==3.0.1
numpy==1.26.4
Pillow==12.1.1
scikit-learn==1.8.0
tqdm==4.67.3
pymysql==1.1.2
cryptography==46.0.5
requests==2.33.0
gradio_client
